+++ server/sessions/internal/controller/sessions/controller.go
@@ -9,11 +9,11 @@ import (
 )
 
 type Repository interface {
-	Save(ctx context.Context, userID int32, token string, expiresUTCNano int64) (err error)
+	Save(ctx context.Context, userID int64, token string, expiresUTCNano int64) (err error)
 	Get(ctx context.Context, token string) (session *model.Session, err error)
-	List(ctx context.Context, userID int32) (sessions []*model.Session, err error)
+	List(ctx context.Context, userID int64) (sessions []*model.Session, err error)
 	Delete(ctx context.Context, token string) (err error)
-	DeleteAll(ctx context.Context, userID int32) (err error)
+	DeleteAll(ctx context.Context, userID int64) (err error)
 }
 
 type Controller struct {
@@ -24,7 +24,7 @@ func New(repo Repository) *Controller {
 	return &Controller{repo}
 }
 
-func (c *Controller) CreateSession(ctx context.Context, userID int32) (session *model.Session, err error) {
+func (c *Controller) CreateSession(ctx context.Context, userID int64) (session *model.Session, err error) {
 	token := utils.RandomString(10)
 	expiresUTCNano := time.Now().Add(time.Hour * 24 * 5).UnixNano()
 
@@ -46,7 +46,7 @@ func (c *Controller) GetSession(ctx context.Context, token string) (*model.Sessi
 	return c.repo.Get(ctx, token)
 }
 
-func (c *Controller) ListUserSessions(ctx context.Context, userID int32) ([]*model.Session, error) {
+func (c *Controller) ListUserSessions(ctx context.Context, userID int64) ([]*model.Session, error) {
 	return c.repo.List(ctx, userID)
 }
 
@@ -54,6 +54,6 @@ func (c *Controller) RemoveSession(ctx context.Context, token string) error {
 	return c.repo.Delete(ctx, token)
 }
 
-func (c *Controller) RemoveUserSessions(ctx context.Context, userID int32) error {
+func (c *Controller) RemoveUserSessions(ctx context.Context, userID int64) error {
 	return c.repo.DeleteAll(ctx, userID)
 }
